don't update skin thumbnails in migration since they are stored in the gem
