+++ card/mod/bootstrap/script/update_skin_thumbnails.rb
@@ -1,0 +1,10 @@
+# -*- encoding : utf-8 -*-
+
+require_relative "../db/migrate_core_cards/lib/skin"
+
+puts "Updating bootstrap themes ..."
+Skin.themes.each do |theme_name|
+  puts theme_name
+  Skin.new(theme_name).update_thumbnail
+end
+
